requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
Flask==2.3.2
Flask-JWT==0.3.2
Flask-Session==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pipreqs==0.4.13
PyJWT==1.4.2
requests==2.31.0
urllib3==2.0.2
Werkzeug==2.3.4
yarg==0.1.9
